+++ Tests/TuistSupportIntegrationTests/System/SystemIntegrationTests.swift
@@ -1,5 +1,6 @@
 import RxBlocking
 import RxSwift
+import Basic
 import XCTest
 @testable import TuistSupport
 @testable import TuistSupportTesting
@@ -45,5 +46,34 @@ final class SystemIntegrationTests: TuistTestCase {
             XCTAssertEqual(elements.count, 0)
             XCTFail("Expected command not to fail but failed with error: \(error)")
         }
+    }
+    
+    func test_pass_DEVELOPER_DIR() throws {
+        
+        try sandbox("DEVELOPER_DIR", value: "/Applications/Xcode/Xcode-10.2.1.app/Contents/Developer/") {
+            let result = try subject.capture("env")
+            XCTAssertTrue(result.contains("DEVELOPER_DIR"))
+        }
+
+    }
+    
+    func test_without_DEVELOPER_DIR() throws {
+        let result = try subject.capture("env")
+        XCTAssertFalse(result.contains("DEVELOPER_DIR"))
+    }
+    
+    func test_do_not_pass_SECRET_VARIABLE() throws {
+        
+        try sandbox("SECRET_VARIABLE", value: "password") {
+            let result = try subject.capture("env")
+            XCTAssertFalse(result.contains("SECRET_VARIABLE"))
+        }
+
+    }
+    
+    func sandbox(_ name: String, value: String, do block: () throws -> Void) rethrows {
+        try? ProcessEnv.setVar(name, value: value)
+        _ = try? block()
+        try? ProcessEnv.unsetVar(name)
     }
 }
